src: Gaurd against use after close.
The leveldown C++ binding segfaults on use after close/free.

Here in the JS layer we add `assert()` checks before every single
API call into the C++ as well as at the start of some methods.

This turns a segfault into an uncaught exception or unhandled
rejection which is significantly more friendly to the user then
having the process abort & dump a core file.

r: @heapwolf