src: Add support for getCache()
When using leveldown in an electron renderer process you
will want to expose leveldown via a contextBridge. The context
bridge exposes a bunch of proxy methods and not objects directly.

The only way to access the cache field is via a function call
that goes into the preload script with contextBridge.